backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Environment and utilities
python-dotenv==1.0.0

# Vector Database
chromadb==0.4.22

# Firebase Authentication
firebase-admin==6.4.0

# HTTP Client
httpx==0.26.0

# ML/Embeddings (fallback)
numpy==1.24.3
sentence-transformers==2.2.2

# Utilities
python-multipart==0.0.6
